+++ src/libbluray/decoders/graphics_controller.h
@@ -25,6 +25,7 @@
 #include "libbluray/bdnav/uo_mask_table.h"
 
 #include <stdint.h>
+#include <stddef.h> /* size_t */
 
 /*
  *
@@ -131,6 +132,6 @@ BD_PRIVATE int                  gc_run(GRAPHICS_CONTROLLER *p,
  */
 
 BD_PRIVATE int                  gc_add_font(GRAPHICS_CONTROLLER *p,
-                                            void *data, int64_t size);
+                                            void *data, size_t size);
 
 #endif // _GRAPHICS_CONTROLLER_H_
